Reorganized pipeline files
