+++ xtrack/pipeline/__init__.py
@@ -1,0 +1,3 @@
+from .core import PipelineStatus, PipelineID
+from .multitracker import PipelineMultiTracker, PipelineBranch
+from .manager import PipelineManager
