+++ xtrack/pipeline/manager.py
@@ -1,13 +1,3 @@
-class PipelineStatus:
-    def __init__(self, on_hold, data=None):
-        self.on_hold = on_hold
-        self.data = data
-
-class PipelineID:
-    def __init__(self,rank,number=0):
-        self.number = number
-        self.rank = rank
-
 class PipelineManager:
     def __init__(self,communicator):
         self._IDs = {}
@@ -66,25 +56,3 @@ class PipelineManager:
     def recieve_message(self,recieve_buffer,element_name,sender_name,reciever_name,internal_tag=0):
         tag = self.get_message_tag(element_name=element_name,sender_name=sender_name,reciever_name=reciever_name,internal_tag=internal_tag)
         self._communicator.Recv(recieve_buffer,source=self.get_particles_rank(sender_name),tag=tag)
-
-class PipelineBranch:
-    def __init__(self, tracker, particles):
-        self.tracker = tracker
-        self.particles = particles
-        self.pipeline_status = None
-
-        self.tracker.enable_pipeline_hold = True
-
-class PipelineMultiTracker:
-
-    def __init__(self, branches):
-        self.branches = branches
-
-    def track(self, **kwargs):
-        for branch in self.branches:
-            branch.pipeline_status = branch.tracker.track(
-                 branch.particles, **kwargs)
-
-        while True:
-            prrrrrr
-
